requirements.txt
# Core dependencies
python-dotenv==1.0.0
pyodbc==4.0.39
pymssql==2.2.8
redis==4.5.5
requests==2.31.0
pytz==2023.3
pydantic==2.4.2

# Web framework for OAuth callback
flask==2.3.3

# Scheduling
schedule==1.2.0

# Logging
loguru==0.7.0

# Database
sqlalchemy==2.0.20

# Testing
pytest==7.4.0
pytest-asyncio==0.21.1
pytest-mock==3.11.1

# Type hints
typing-extensions==4.7.1

# Date handling
python-dateutil==2.8.2

# Rate limiting
ratelimit==2.2.1

# OAuth handling
oauthlib==3.2.2
requests-oauthlib==1.3.1 